Update Builder.php
strtotime() evaluates to negative timestamp for some strings (ex.: 0063A0000181Fpl, 0063A0000181Fby) and this part of the "elseif" does not get executed, so in case someone wants to send in an ID using the soqlBuilder, it will get an error from Salesforce, because Salesforce only accepts ID's as strings. If this "elseif" part (line 75) does not evaluate to true, teh string won't get wrapped in single quotes and the request fails.